The Kinect wrapper class now sends out Bitmap Ready events after when the MultiSourceFeed is updated, FrameReader and KinectSensor are moved out of StandbyWindow and into the Kinect class
