+++ gbtis/StandbyWindow.xaml.cs
@@ -1,18 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Timers;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
 using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-using Microsoft.Kinect;
 
 namespace gbtis {
     /// <summary>
@@ -27,7 +17,6 @@ namespace gbtis {
 
         // For the kinect display
         Kinect kinect;
-        MultiSourceFrameReader frameReader;
 
         /// <summary>
         /// Constructor for the standby window
@@ -42,9 +31,7 @@ namespace gbtis {
                 gbtis.Properties.Resources.msgStart :
                 gbtis.Properties.Resources.msgNoSensor;
 
-            // Prepare sensor feed
-            frameReader = kinect.sensor.OpenMultiSourceFrameReader(FrameSourceTypes.Color);
-            frameReader.MultiSourceFrameArrived += FrameReader_MultiSourceFrameArrived;
+            kinect.BitMapReady += BitMapArrived;
 
             // Initialize the names
             topName.Text = "";
@@ -62,15 +49,10 @@ namespace gbtis {
         /// Update the camera feed from the sensor
         /// </summary>
         /// <param name="sender">Object sending the event</param>
-        /// <param name="e">Arguments for the event object</param>
-        void FrameReader_MultiSourceFrameArrived(object sender, MultiSourceFrameArrivedEventArgs e) {
-            var reference = e.FrameReference.AcquireFrame();
-            using (var frame = reference.ColorFrameReference.AcquireFrame()) {
-                if (frame != null) {
-                    this.Dispatcher.Invoke(new Action(() =>
-                    sensorFeed.Source = ToBitmap(frame)));
-                }
-            }
+        /// <param name="e">The latest ImageSource</param>
+        void BitMapArrived(object sender, ImageSource img) {
+            this.Dispatcher.Invoke(new Action(() =>
+                sensorFeed.Source = img));
         }
 
         /// <summary>
@@ -88,29 +70,6 @@ namespace gbtis {
                     new TextBlockFadeAnimation(bottomName, names[(i + 2) % names.Count()]);
                 });
             } catch (OperationCanceledException) { }
-        }
-
-        /// <summary>
-        /// Convert a frame of kinect color video to bitmap for display
-        /// Conversion code from http://pterneas.com/2014/02/20/kinect-for-windows-version-2-color-depth-and-infrared-streams/
-        /// Under Apache License 2.0
-        /// </summary>
-        /// <param name="frame">The frame received from the kinect</param>
-        /// <returns>A bitmap format source for a WPF image control</returns>
-        private static ImageSource ToBitmap(ColorFrame frame) {
-            int width = frame.FrameDescription.Width;
-            int height = frame.FrameDescription.Height;
-            
-            byte[] pixels = new byte[width * height * ((PixelFormats.Bgr32.BitsPerPixel + 7) / 8)];
-
-            if (frame.RawColorImageFormat == ColorImageFormat.Bgra) {
-                frame.CopyRawFrameDataToArray(pixels);
-            } else {
-                frame.CopyConvertedFrameDataToArray(pixels, ColorImageFormat.Bgra);
-            }
-
-            int stride = width * PixelFormats.Bgr32.BitsPerPixel / 8;
-            return BitmapSource.Create(width, height, 96, 96, PixelFormats.Bgr32, null, pixels, stride);
         }
 
         /// <summary>
